Improve job status display and error handling with shared utilities
Refactor job status display and error handling by moving shared functions to `src/utils/jobHelpers.tsx`, adding date validation, improving `cancelJob` error catching, and setting `staleTime` for job stats.
